fully reformatted task 2 of homework 13
