--- homework/flask_alembic_hw/al_api_hotel/alembic/versions/3208d748d4ee_first_commit.py
@@ -1,52 +1,0 @@
-"""First commit
-
-Revision ID: 3208d748d4ee
-Revises: 
-Create Date: 2021-03-13 10:01:31.725157
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '3208d748d4ee'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('owner',
-    sa.Column('id', sa.INTEGER(), nullable=False),
-    sa.Column('name', sa.String(length=50), nullable=True),
-    sa.Column('phone', sa.String(length=50), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('phone')
-    )
-    op.create_table('pet',
-    sa.Column('id', sa.INTEGER(), nullable=False),
-    sa.Column('name', sa.String(length=50), nullable=True),
-    sa.Column('owner_id', sa.INTEGER(), nullable=False),
-    sa.Column('room_num', sa.INTEGER(), nullable=True),
-    sa.ForeignKeyConstraint(['owner_id'], ['owner.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('activity',
-    sa.Column('id', sa.INTEGER(), nullable=False),
-    sa.Column('pet_id', sa.INTEGER(), nullable=False),
-    sa.Column('activity_type', sa.String(length=50), nullable=True),
-    sa.Column('time', sa.Time(), nullable=True),
-    sa.ForeignKeyConstraint(['pet_id'], ['pet.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('activity')
-    op.drop_table('pet')
-    op.drop_table('owner')
-    # ### end Alembic commands ###
