+++ homework/flask_alembic_hw/al_api_hotel/alembic/versions/8070904a4954_initial_revision.py
@@ -1,0 +1,34 @@
+"""Initial revision
+
+Revision ID: 8070904a4954
+Revises: 
+Create Date: 2021-03-18 17:17:32.389078
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '8070904a4954'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('owner',
+    sa.Column('id', sa.INTEGER(), nullable=False),
+    sa.Column('name', sa.String(length=50), nullable=True),
+    sa.Column('phone', sa.String(length=50), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('phone')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('owner')
+    # ### end Alembic commands ###
