+++ homework/flask_alembic_hw/al_api_hotel/models.py
@@ -15,21 +15,3 @@ class Owner(Base):
     id = Column(INTEGER, primary_key=True)
     name = Column(String(50))
     phone = Column(String(50), unique=True)
-
-
-class Pet(Base):
-    __tablename__ = 'pet'
-
-    id = Column(INTEGER, primary_key=True)
-    name = Column(String(50))
-    owner_id = Column(INTEGER, ForeignKey('owner.id'), nullable=False)
-    room_num = Column(INTEGER)
-
-
-class Activity(Base):
-    __tablename__ = 'activity'
-
-    id = Column(INTEGER, primary_key=True)
-    pet_id = Column(INTEGER, ForeignKey('pet.id'), nullable=False)
-    activity_type = Column(String(50))
-    time = Column(Time)
